Reimplementing AnswerStateImpl from AnswerState interfaces
