+++ reasoner/resolution/answer/AnswerState.java
@@ -1,23 +1,5 @@
-/*
- * Copyright (C) 2021 Grakn Labs
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
 package grakn.core.reasoner.resolution.answer;
 
-import grakn.common.collection.Pair;
 import grakn.core.common.exception.GraknException;
 import grakn.core.concept.Concept;
 import grakn.core.concept.answer.ConceptMap;
@@ -25,1207 +7,397 @@ import grakn.core.concept.answer.ExplainableAnswer;
 import grakn.core.concurrent.actor.Actor;
 import grakn.core.logic.Rule;
 import grakn.core.logic.resolvable.Unifier;
-import grakn.core.logic.resolvable.Unifier.Requirements.Instance;
 import grakn.core.pattern.Conjunction;
 import grakn.core.reasoner.resolution.framework.Resolver;
-import grakn.core.reasoner.resolution.resolver.RootResolver;
 import grakn.core.traversal.common.Identifier;
 
-import javax.annotation.Nullable;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 
-import static grakn.common.collection.Collections.set;
 import static grakn.common.util.Objects.className;
 import static grakn.core.common.exception.ErrorMessage.Internal.ILLEGAL_CAST;
-import static grakn.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
 import static grakn.core.common.exception.ErrorMessage.Pattern.INVALID_CASTING;
-import static grakn.core.common.iterator.Iterators.iterate;
 
-public abstract class AnswerState {
+public interface AnswerState {
 
-    private final Actor.Driver<? extends Resolver<?>> root;
-    final ConceptMap conceptMap;
-    final boolean requiresReiteration;
+    interface Explainable {
 
-    AnswerState(ConceptMap conceptMap, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-        this.conceptMap = conceptMap;
-        this.root = root;
-        this.requiresReiteration = requiresReiteration;
-    }
-
-    public abstract ConceptMap conceptMap();
-
-    public boolean requiresReiteration() {
-        return requiresReiteration;
-    }
-
-    public boolean isTop() { return false; }
-
-    public Top asTop() {
-        throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Top.class));
-    }
-
-    public boolean isPartial() { return false; }
-
-    public Partial<?, ?> asPartial() {
-        throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Partial.Compound.class));
-    }
-
-    public Actor.Driver<? extends Resolver<?>> root() {
-        return root;
-    }
-
-    public static abstract class Top2 extends AnswerState {
-
-        public Top2(ConceptMap conceptMap, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-            super(conceptMap, root, requiresReiteration);
-        }
-
-        public boolean isTop() { return true; }
-
-        public Top2 asTop() {
-            return this;
-        }
-
-        public boolean isMatch() {
-            return false;
-        }
-
-        public Match.Top asMatch() {
-            throw GraknException.of(ILLEGAL_CAST, this.getClass(), Top.Match.class);
-        }
-
-        public boolean isExplain() {
-            return false;
-        }
-
-        public Explain.Top asExplain() {
-            throw GraknException.of(ILLEGAL_CAST, this.getClass(), Top.Explain.class);
-        }
+        boolean explainable();
 
     }
 
-    public static abstract class Partial2 extends AnswerState {
+    ConceptMap conceptMap();
 
-    }
+    boolean requiresReiteration();
 
-    public static abstract class Match extends AnswerState {
+    Actor.Driver<? extends Resolver<?>> root();
 
+    default boolean isTop() { return false; }
 
-        public static abstract class Top extends Top2 {
+    default Top asTop() { throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Top.class)); }
 
-            final boolean explainable;
-            final Set<Identifier.Variable.Name> getFilter;
-            private final int hash;
+    default boolean isPartial() { return false; }
 
-            Top(ConceptMap conceptMap, @Nullable Set<Identifier.Variable.Name> getFilter, Actor.Driver<? extends Resolver<?>> root,
-                  boolean requiresReiteration, boolean explainable) {
-                super(conceptMap, root, requiresReiteration);
-                this.getFilter = getFilter;
-                this.explainable = explainable;
-                this.hash = Objects.hash(conceptMap, getFilter, root, requiresReiteration, explainable);
-            }
+    default Partial<?, ?> asPartial() { throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Partial.class)); }
 
-            public static Match.Top.Initial initial(Set<Identifier.Variable.Name> getFilter, Actor.Driver<? extends Resolver<?>> root, boolean explainable) {
-                return new Match.Top.Initial(new ConceptMap(), getFilter, root, explainable);
-            }
+    interface Top extends AnswerState {
+
+        default boolean isTop() { return true; }
+
+        default Top asTop() { return this; }
+
+        default boolean isMatch() { return false; }
+
+        default Match asMatch() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Match.class); }
+
+        default boolean isExplain() { return false; }
+
+        default Explain asExplain() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Explain.class); }
+
+        interface Match extends Top, Explainable {
+
+            Set<Identifier.Variable.Name> getFilter();
 
             @Override
-            public ConceptMap conceptMap() {
-                return conceptMap.filter(getFilter);
-            }
-
-            @Override
-            public boolean isMatch() {
+            default boolean isMatch() {
                 return true;
             }
 
             @Override
-            public Match.Top asMatch() {
+            default Match asMatch() {
                 return this;
             }
 
-            public boolean isFinished() {
-                return false;
+            default boolean isFinished() { return false; }
+
+            default Finished asFinished() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Finished.class); }
+
+            interface Initial extends Match {
+
+                Partial.Compound.Match.Root toDownstream();
+
+                Finished finish(ConceptMap conceptMap, boolean requiresReiteration);
+
             }
 
-            public Top.Match.Finished asFinished() {
-                throw GraknException.of(ILLEGAL_CAST, this.getClass(), Top.Match.Finished.class);
-            }
-
-            @Override
-            public boolean equals(Object o) {
-                if (this == o) return true;
-                if (o == null || getClass() != o.getClass()) return false;
-                Match.Top.Initial that = (Match.Top.Initial) o;
-                return Objects.equals(root(), that.root()) &&
-                        Objects.equals(conceptMap, that.conceptMap) &&
-                        Objects.equals(getFilter, that.getFilter) &&
-                        requiresReiteration == that.requiresReiteration &&
-                        explainable == that.explainable;
-            }
-
-            @Override
-            public int hashCode() {
-                return hash;
-            }
-
-            public static class Initial extends Match.Top {
-
-                Initial(ConceptMap conceptMap, @Nullable Set<Identifier.Variable.Name> getFilter, Actor.Driver<? extends Resolver<?>> root, boolean explain) {
-                    super(conceptMap, getFilter, root, false, explain);
-                }
-
-                public Match.Partial.Compound.Root toDownstream() {
-                    return Match.Partial.Compound.Root.create(conceptMap, this, root(), explainable);
-                }
-
-                Match.Top.Finished finish(ConceptMap conceptMap, boolean requiresReiteration) {
-                    return new Match.Top.Finished(conceptMap, getFilter, root(), requiresReiteration, explainable);
-                }
+            interface Finished extends Match {
 
                 @Override
-                public String toString() {
-                    return "AnswerState.Match.Top.Initial{" +
-                            "root=" + root() +
-                            ", conceptMap=" + conceptMap +
-                            ", filter=" + getFilter +
-                            '}';
-                }
-
-            }
-
-            public static class Finished extends Match.Top {
-
-                Finished(ConceptMap conceptMap, @Nullable Set<Identifier.Variable.Name> getFilter, Actor.Driver<? extends Resolver<?>> root,
-                         boolean requiresReiteration, boolean explainable) {
-                    super(conceptMap, getFilter, root, requiresReiteration, explainable); }
+                default boolean isFinished() { return true; }
 
                 @Override
-                public boolean isFinished() {
-                    return true;
-                }
-
-                @Override
-                public Match.Top.Finished asFinished() {
-                    return this;
-                }
-
-                @Override
-                public String toString() {
-                    return "AnswerState.Match.Top.Finished{" +
-                            "root=" + root() +
-                            ", conceptMap=" + conceptMap +
-                            ", filter=" + getFilter +
-                            ", requiresReiteration=" + requiresReiteration +
-                            '}';
-                }
-            }
-        }
-
-        public static class Partial extends Partial2 {
-
-            public static abstract class Compound extends Partial {
-
-                public static class Root extends Compound {
-
-                }
-
-                public static class NonRoot extends Compound {
-
-                }
-
-                public static class Condition extends Compound {
-
-                }
-
-            }
-
-            public static class Concludable extends Partial {
-
-            }
-
-            public static class Conclusion extends Partial {
+                default Finished asFinished() { return this; }
 
             }
 
         }
 
+        interface Explain extends Top {
+
+            @Override
+            default boolean isExplain() { return true; }
+
+            @Override
+            default Explain asExplain() { return this; }
+
+            default boolean isFinished() { return false; }
+
+            default Finished asFinished() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Finished.class); }
+
+            interface Initial extends Explain {
+
+                Partial.Compound.Explain.Root toDownstream();
+
+                Finished finish(ConceptMap conceptMap, boolean requiresReiteration, Explanation explanation);
+
+            }
+
+            interface Finished extends Explain {
+
+                Explanation explanation();
+
+                @Override
+                default boolean isFinished() { return true; }
+
+                @Override
+                default Finished asFinished() { return this; }
+
+            }
+        }
     }
 
-    public static abstract class Explain extends AnswerState {
+    interface Partial<SELF extends Partial<SELF, PARENT>, PARENT extends AnswerState> extends AnswerState {
 
-        public static class Top extends Top2 {
-
-            public Top(ConceptMap conceptMap, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-                super(conceptMap, root, requiresReiteration);
-            }
-
-            public static Explain.Top.Initial initial(ConceptMap bounds, Actor.Driver<RootResolver.Explain> root) {
-                return new Explain.Top.Initial(bounds, root, false);
-            }
-
-            @Override
-            public ConceptMap conceptMap() {
-                return conceptMap;
-            }
-
-            @Override
-            public boolean isExplain() {
-                return true;
-            }
-
-            @Override
-            public Explain.Top asExplain() {
-                return this;
-            }
-
-            public boolean isFinished() {
-                return false;
-            }
-
-            public Explain.Top.Finished asFinished() {
-                throw GraknException.of(ILLEGAL_CAST, this.getClass(), Explain.class);
-            }
-
-            public static class Initial extends Explain.Top {
-
-                private final int hash;
-
-                public Initial(ConceptMap conceptMap, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-                    super(conceptMap, root, requiresReiteration);
-                    this.hash = Objects.hash(root, conceptMap, requiresReiteration);
-                }
-
-                public Explain.Partial.Compound.Root toDownstream() {
-                    return Explain.Partial.Compound.root(conceptMap, this, root());
-                }
-
-                public Finished finish(ConceptMap conceptMap, boolean requiresReiteration, Explanation explanation) {
-                    return new Finished(conceptMap, root(), requiresReiteration, explanation);
-                }
-
-                @Override
-                public boolean equals(Object o) {
-                    if (this == o) return true;
-                    if (o == null || getClass() != o.getClass()) return false;
-                    AnswerState.Explain.Top.Initial that = (AnswerState.Explain.Top.Initial) o;
-                    return Objects.equals(root(), that.root()) &&
-                            Objects.equals(conceptMap, that.conceptMap) &&
-                            requiresReiteration == that.requiresReiteration;
-                }
-
-                @Override
-                public int hashCode() {
-                    return hash;
-                }
-            }
-
-            public static class Finished extends Explain.Top {
-
-                private final Explanation explanation;
-                private final int hash;
-
-                public Finished(ConceptMap conceptMap, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration,
-                                Explanation explanation) {
-                    super(conceptMap, root, requiresReiteration);
-                    this.explanation = explanation;
-                    this.hash = Objects.hash(root, conceptMap, requiresReiteration, explanation);
-                }
-
-                @Override
-                public boolean isFinished() {
-                    return true;
-                }
-
-                @Override
-                public Finished asFinished() {
-                    return this;
-                }
-
-                public Explanation explanation() {
-                    return explanation;
-                }
-
-                @Override
-                public String toString() {
-                    return "AnswerState.Explain.To.Finished{" +
-                            "root=" + root() +
-                            ", conceptMap=" + conceptMap +
-                            ", requiresReiteration=" + requiresReiteration +
-                            ", explanation=" + explanation +
-                            '}';
-                }
-
-                @Override
-                public boolean equals(Object o) {
-                    if (this == o) return true;
-                    if (o == null || getClass() != o.getClass()) return false;
-                    Explain.Top.Finished that = (Explain.Top.Finished) o;
-                    return Objects.equals(root(), that.root()) &&
-                            Objects.equals(conceptMap, that.conceptMap) &&
-                            requiresReiteration == that.requiresReiteration &&
-                            Objects.equals(explanation, that.explanation);
-                }
-
-                @Override
-                public int hashCode() {
-                    return hash;
-                }
-
-            }
-        }
-
-        public static class Partial extends Partial2 {
-
-            public static abstract class Compound extends Partial {
-
-                public static class Root extends Compound {
-
-                }
-
-                public static class Condition extends Compound {
-
-                }
-
-            }
-
-            public static class Concludable extends Partial {
-
-            }
-
-            public static class Conclusion extends Partial {
-
-            }
-
-        }
-
-    }
-
-    public static abstract class Partial<SELF extends Partial<SELF, PARENT>, PARENT extends AnswerState> extends AnswerState {
-
-        final PARENT parent;
-
-        public Partial(ConceptMap partialAnswer, PARENT parent, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-            super(partialAnswer, root, requiresReiteration);
-            this.parent = parent;
-        }
-
-        public ConceptMap conceptMap() {
-            return conceptMap;
-        }
-
-        protected PARENT parent() {
-            return parent;
-        }
+        PARENT parent();
 
         @Override
-        public boolean isPartial() { return true; }
+        default boolean isPartial() { return true; }
 
         @Override
-        public Partial<?, ?> asPartial() {
-            return this;
-        }
+        default Partial<?, ?> asPartial() { return this; }
 
-        public boolean isCompound() { return false; }
+        default boolean isCompound() { return false; }
 
-        public boolean isConcludable() { return false; }
+        default boolean isConcludable() { return false; }
 
-        public boolean isConclusion() { return false; }
+        default boolean isConclusion() { return false; }
 
-        public Compound<?, ?> asCompound() {
+        default Compound<?, ?> asCompound() {
             throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Compound.class));
         }
 
-        public Concludable<?, ?> asConcludable() {
+        default Concludable<?, ?> asConcludable() {
             throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Concludable.class));
         }
 
-        public Conclusion<?, ?> asConclusion() {
-            throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Conclusion.Match.class));
+        default Conclusion<?, ?> asConclusion() {
+            throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Conclusion.class));
         }
 
-        protected ConceptMap extendAnswer(ConceptMap extension) {
-            return extendAnswer(extension, null);
-        }
+        interface Compound<SLF extends Compound<SLF, PRNT>, PRNT extends AnswerState> extends Partial<SLF, PRNT> {
 
-        protected ConceptMap extendAnswer(ConceptMap extension, ExplainableAnswer explainableAnswer) {
-            /*
-            We MUST retain initial concepts, and add derived answers afterward. It's possible, and correct,
-            that the derived answers overlap but are different: for example, when a subtype is found
-            by the derived answer, but the initial already uses the supertype.
-             */
-            Map<Identifier.Variable.Retrievable, Concept> concepts = new HashMap<>(extension.concepts());
-            // add the initial concept map second, to make sure we override and retain all of these
-            concepts.putAll(conceptMap().concepts());
-            return new ConceptMap(concepts, explainableAnswer);
-        }
-
-        abstract SELF with(ConceptMap extension, boolean requiresReiteration);
-
-        public static abstract class Compound<SLF extends Compound<SLF, PRNT>, PRNT extends AnswerState>
-                extends Partial<SLF, PRNT> {
-
-            public Compound(ConceptMap partialAnswer, PRNT parent, Actor.Driver<? extends Resolver<?>> root,
-                            boolean requiresReiteration) {
-                super(partialAnswer, parent, root, requiresReiteration);
-            }
-
-            public Match.NonRoot filterToDownstream(Set<Identifier.Variable.Retrievable> filter) {
-                return Match.NonRoot.create(this, filter, root());
-            }
-
-            public abstract Concludable<?, ?> mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction);
+            Concludable<?, ?> mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction);
 
             @Override
-            public boolean isCompound() { return true; }
+            default boolean isCompound() { return true; }
 
             @Override
-            public Compound<?, ?> asCompound() { return this; }
+            default Compound<?, ?> asCompound() { return this; }
 
-            public boolean isRoot() { return false; }
+            default boolean isMatch() { return false; }
 
-            public Match.Root asRoot() {
-                throw GraknException.of(ILLEGAL_CAST, this.getClass(), Match.Root.class);
-            }
+            default Match<?, ?> asMatch() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Match.class); }
 
-            public boolean isExplainRoot() { return false; }
+            default boolean isExplain() { return false; }
 
-            public ExplainRoot asExplainRoot() {
-                throw GraknException.of(ILLEGAL_CAST, this.getClass(), Match.Root.class);
-            }
+            default Explain<?, ?> asExplain() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Explain.class); }
 
-            public boolean isCondition() { return false; }
+            interface Match<S extends Match<S, P>, P extends AnswerState> extends Compound<S, P>, Explainable {
 
-            public Match.Condition<?> asCondition() {
-                throw GraknException.of(ILLEGAL_CAST, this.getClass(), Match.Condition.class);
-            }
+                // TODo does everyone have to implement this? It's used for receiving from a NonRoot
+                S with(ConceptMap extension, boolean requiresReiteration);
 
-            public boolean isNonRoot() { return false; }
+                S with(ConceptMap extension, boolean requiresReiteration, Conjunction source);
 
-            public Match.NonRoot asNonRoot() {
-                throw GraknException.of(ILLEGAL_CAST, this.getClass(), Match.NonRoot.class);
-            }
+                default boolean isRoot() { return false; }
 
-            public static abstract class Match<S extends Match<S, P>, P extends AnswerState> extends Compound<S, P> {
+                default Root asRoot() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Root.class); }
 
-                public Match(ConceptMap partialAnswer, P parent, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-                    super(partialAnswer, parent, root, requiresReiteration);
-                }
+                default boolean isCondition() { return false; }
 
-                abstract S with(ConceptMap extension, boolean requiresReiteration, Conjunction source);
+                default Condition<?> asCondition() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Condition.class); }
 
-                public static class Root extends Match<Root, Top.Match.Initial> {
+                default boolean isNonRoot() { return false; }
 
-                    private final boolean explainable;
-                    private final Set<ExplainableAnswer.Explainable> explainables;
-                    private final int hash;
+                default NonRoot asNonRoot() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), NonRoot.class); }
 
-                    public Root(ConceptMap partialAnswer, Top.Match.Initial parent, Actor.Driver<? extends Resolver<?>> root,
-                                boolean requiresReiteration, boolean explainable, Set<ExplainableAnswer.Explainable> explainables) {
-                        super(partialAnswer, parent, root, requiresReiteration);
-                        this.explainable = explainable;
-                        this.explainables = explainables;
-                        this.hash = Objects.hash(root, conceptMap, parent, explainable);
-                    }
-
-                    static Root create(ConceptMap conceptMap, Top.Match.Initial parent, Actor.Driver<? extends Resolver<?>> root, boolean explain) {
-                        return new Root(conceptMap, parent, root, false, explain, explain ? new HashSet<>() : null);
-                    }
+                interface Root extends Match<Root, Top.Match.Initial> {
 
                     @Override
-                    public boolean isRoot() { return true; }
+                    Concludable.Match<Root> mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction);
+
+                    Top.Match.Finished toFinishedTop(Conjunction conjunctionAnswered);
 
                     @Override
-                    public Root asRoot() { return this; }
+                    default boolean isRoot() { return true; }
 
                     @Override
-                    Root with(ConceptMap extension, boolean requiresReiteration) {
-                        return new Root(extendAnswer(extension), parent(), root(), requiresReiteration, explainable, explainables);
-                    }
-
-                    @Override
-                    public Root with(ConceptMap extension, boolean requiresReiteration, Conjunction source) {
-                        Set<ExplainableAnswer.Explainable> explainablesExtended;
-                        if (explainable) {
-                            explainablesExtended = new HashSet<>(explainables);
-                            explainablesExtended.add(ExplainableAnswer.Explainable.unidentified(source));
-                        } else {
-                            explainablesExtended = null;
-                        }
-                        return new Root(extendAnswer(extension), parent(), root(), requiresReiteration, explainable, explainablesExtended);
-                    }
-
-                    @Override
-                    public Concludable.Match<Root> mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction) {
-                        return Concludable.match(this, mapping, root(), nextResolverConjunction, explainable);
-                    }
-
-                    public Top.Match.Finished toFinishedTop(Conjunction conjunctionAnswered) {
-                        // TODO add the conjunction and explainables into the concept map
-                        return parent().finish(conceptMap(), requiresReiteration);
-                    }
-
-                    @Override
-                    public String toString() {
-                        return "AnswerState.Partial.Compound.Root{" +
-                                "root=" + root() +
-                                ", conceptMap=" + conceptMap() +
-                                ", explainables=" + explainables +
-                                '}';
-                    }
-
-                    @Override
-                    public boolean equals(Object o) {
-                        if (this == o) return true;
-                        if (o == null || getClass() != o.getClass()) return false;
-                        Root that = (Root) o;
-                        return Objects.equals(root(), that.root()) &&
-                                Objects.equals(conceptMap, that.conceptMap) &&
-                                Objects.equals(parent, that.parent) &&
-                                explainable == that.explainable;
-                    }
-
-                    @Override
-                    public int hashCode() {
-                        return hash;
-                    }
+                    default Root asRoot() { return this; }
 
                 }
 
-                public static class Condition<P extends Conclusion<P, ?>> extends Match<Condition<P>, P> {
+                interface NonRoot extends Match<NonRoot, Compound.Match<?, ?>> {
 
-                    private final Set<ExplainableAnswer.Explainable> explainables;
-                    private final Set<Identifier.Variable.Retrievable> filter;
-                    private final int hash;
-
-                    private Condition(ConceptMap filteredMap, P parent, Set<Identifier.Variable.Retrievable> filter,
-                                      Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration, Set<ExplainableAnswer.Explainable> explainables) {
-                        super(filteredMap, parent, root, requiresReiteration);
-                        this.explainables = explainables;
-                        this.filter = filter;
-                        this.hash = Objects.hash(root, conceptMap, parent, filter);
-                    }
-
-                    static <P extends Conclusion<P, ?>> Condition<P> create(P parent, Set<Identifier.Variable.Retrievable> filter,
-                                                                            Actor.Driver<? extends Resolver<?>> root) {
-                        return new Condition<>(parent.conceptMap().filter(filter), parent, filter, root, false, set());
-                    }
+                    Set<Identifier.Variable.Retrievable> filter();
 
                     @Override
-                    public Concludable.Match<Condition<P>> mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction) {
-                        return Concludable.match(this, mapping, root(), nextResolverConjunction, false);
-                    }
+                    Concludable.Match<NonRoot> mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction);
 
-                    public P toUpstream(Conjunction conditionConjunction) {
-                        if (conceptMap().concepts().isEmpty()) throw GraknException.of(ILLEGAL_STATE);
-                        // TODO we don't want explainables all, we only need them in `ExplainCondition`
-                        ExplainableAnswer conditionAnswer = new ExplainableAnswer(conceptMap, conditionConjunction, explainables);
-                        return parent().with(conceptMap().filter(filter), requiresReiteration || parent().requiresReiteration(), conditionAnswer);
-                    }
+                    Partial.Compound<?, ?> toUpstream();
+
+                    Partial.Compound<?, ?> aggregateToUpstream(ConceptMap conceptMap);
 
                     @Override
-                    Condition<P> with(ConceptMap extension, boolean requiresReiteration) {
-                        return new Condition<>(extendAnswer(extension), parent(), filter, root(), requiresReiteration, explainables);
-                    }
+                    default boolean isNonRoot() { return true; }
 
                     @Override
-                    public Condition<P> with(ConceptMap extension, boolean requiresReiteration, Conjunction source) {
-                        // TODO we don't want to have explanation or source here at all, we only need them in `ExplainCondition`
-                        Set<ExplainableAnswer.Explainable> explainablesClone = new HashSet<>(explainables);
-                        explainablesClone.add(ExplainableAnswer.Explainable.unidentified(source));
-                        return new Condition<>(extendAnswer(extension), parent(), filter, root(), requiresReiteration, explainablesClone);
-                    }
+                    default NonRoot asNonRoot() { return this; }
+
+                }
+
+                interface Condition extends Match<Condition, Conclusion.Match> {
 
                     @Override
-                    public boolean isCondition() {
+                    Concludable.Match<Condition> mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction);
+
+                    Conclusion.Match toUpstream();
+
+                    @Override
+                    default boolean isCondition() {
                         return true;
                     }
 
                     @Override
-                    public Condition<P> asCondition() {
-                        return this;
-                    }
+                    default Condition asCondition() { return this; }
 
-                    @Override
-                    public String toString() {
-                        return "AnswerState.Partial.Compound.Condition{" +
-                                "root=" + root() +
-                                ", conceptMap=" + conceptMap() +
-                                ", filter=" + filter +
-                                ", explainables=" + explainables +
-                                '}';
-                    }
-
-                    @Override
-                    public boolean equals(Object o) {
-                        if (this == o) return true;
-                        if (o == null || getClass() != o.getClass()) return false;
-                        Condition that = (Condition) o;
-                        return Objects.equals(root(), that.root()) &&
-                                Objects.equals(conceptMap, that.conceptMap) &&
-                                Objects.equals(parent, that.parent) &&
-                                Objects.equals(filter, that.filter);
-                    }
-
-                    @Override
-                    public int hashCode() {
-                        return hash;
-                    }
-                }
-
-                public static class NonRoot extends Match<NonRoot, Compound<?, ?>> {
-
-                    private final Set<Identifier.Variable.Retrievable> filter;
-                    private final int hash;
-
-                    private NonRoot(ConceptMap filteredConceptMap, Compound<?, ?> parent, Set<Identifier.Variable.Retrievable> filter,
-                                    Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-                        super(filteredConceptMap, parent, root, requiresReiteration);
-                        this.filter = filter;
-                        this.hash = Objects.hash(root, conceptMap, parent, filter);
-                    }
-
-                    static NonRoot create(Compound<?, ?> parent, Set<Identifier.Variable.Retrievable> filter,
-                                          Actor.Driver<? extends Resolver<?>> root) {
-                        return new NonRoot(parent.conceptMap().filter(filter), parent, filter, root, false);
-                    }
-
-                    @Override
-                    public boolean isNonRoot() { return true; }
-
-                    @Override
-                    public NonRoot asNonRoot() { return this; }
-
-                    @Override
-                    public Concludable.Match<NonRoot> mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction) {
-                        return Concludable.match(this, mapping, root(), nextResolverConjunction, false);
-                    }
-
-                    public Compound<?, ?> toUpstream() {
-                        if (conceptMap().concepts().isEmpty()) throw GraknException.of(ILLEGAL_STATE);
-                        return parent().with(conceptMap().filter(filter), requiresReiteration || parent().requiresReiteration());
-                    }
-
-                    public Compound<?, ?> aggregateToUpstream(ConceptMap conceptMap) {
-                        if (conceptMap.concepts().isEmpty()) throw GraknException.of(ILLEGAL_STATE);
-                        return parent().with(conceptMap.filter(filter), requiresReiteration || parent().requiresReiteration());
-                    }
-
-                    @Override
-                    NonRoot with(ConceptMap extension, boolean requiresReiteration) {
-                        return new NonRoot(extendAnswer(extension), parent(), filter, root(), requiresReiteration);
-                    }
-
-                    @Override
-                    public NonRoot with(ConceptMap extension, boolean requiresReiteration, Conjunction source) {
-                        return new NonRoot(extendAnswer(extension), parent(), filter, root(), requiresReiteration);
-                    }
-
-                    @Override
-                    public String toString() {
-                        return "AnswerState.Compound.NonRoot{" +
-                                "root=" + root() +
-                                ", conceptMap=" + conceptMap() +
-                                ", filter=" + filter +
-                                '}';
-                    }
-
-                    @Override
-                    public boolean equals(Object o) {
-                        if (this == o) return true;
-                        if (o == null || getClass() != o.getClass()) return false;
-                        NonRoot filtered = (NonRoot) o;
-                        return Objects.equals(root(), filtered.root()) &&
-                                Objects.equals(conceptMap, filtered.conceptMap) &&
-                                Objects.equals(parent, filtered.parent) &&
-                                Objects.equals(filter, filtered.filter);
-                    }
-
-                    @Override
-                    public int hashCode() {
-                        return hash;
-                    }
                 }
             }
 
-            public static class ExplainRoot extends Compound<ExplainRoot, Top.Explain.Initial> {
+            interface Explain<S extends Explain<S, P>, P extends AnswerState> extends Compound<S, P> {
 
-                private final int hash;
-                private final Explanation explanation;
+                default boolean isRoot() { return false; }
 
-                private ExplainRoot(ConceptMap partialAnswer, Top.Explain.Initial parent, Actor.Driver<? extends Resolver<?>> root,
-                                    boolean requiresReiteration, @Nullable Explanation explanation) {
-                    super(partialAnswer, parent, root, requiresReiteration);
-                    this.explanation = explanation;
-                    this.hash = Objects.hash(root, conceptMap, parent, requiresReiteration, explanation);
+                default Root asRoot() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Root.class); }
+
+                default boolean isCondition() { return false; }
+
+                default Condition asCondition() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Condition.class); }
+
+                interface Root extends Explain<Root, Top.Explain.Initial> {
+
+                    Root with(ConceptMap extension, boolean requiresReiteration, Explanation explanation);
+
+                    boolean hasExplanation();
+
+                    @Override
+                    Concludable.Explain mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction);
+
+                    Top.Explain.Finished toFinishedTop();
+
+                    @Override
+                    default boolean isRoot() { return true; }
+
+                    @Override
+                    default Root asRoot() { return this; }
+
                 }
 
-                static ExplainRoot create(ConceptMap conceptMap, Top.Explain.Initial parent, Actor.Driver<? extends Resolver<?>> root) {
-                    return new ExplainRoot(conceptMap, parent, root, false, null);
-                }
+                interface Condition extends Explain<Condition, Conclusion.Explain> {
 
-                @Override
-                ExplainRoot with(ConceptMap extension, boolean requiresReiteration) {
-                    // TODO we don't really want this method here do we?
-                    return new ExplainRoot(extendAnswer(extension), parent(), root(), requiresReiteration, null);
-                }
+                    // TODO
 
-                public ExplainRoot with(ConceptMap extension, boolean requiresReiteration, Explanation explanation) {
-                    return new ExplainRoot(extendAnswer(extension), parent(), root(), requiresReiteration, explanation);
-                }
+                    @Override
+                    default boolean isCondition() { return true; }
 
-                public boolean hasExplanation() {
-                    return explanation != null;
-                }
+                    @Override
+                    default Condition asCondition() { return this; }
 
-                @Override
-                public Concludable.Explain mapToDownstream(Mapping mapping, Conjunction nextResolverConjunction) {
-                    return Concludable.explain(this, mapping, root(), nextResolverConjunction);
-                }
-
-                public Top.Explain.Finished toFinishedTop() {
-                    return parent().finish(conceptMap(), requiresReiteration || parent().requiresReiteration(), explanation);
-                }
-
-                @Override
-                public boolean isExplainRoot() {
-                    return true;
-                }
-
-                @Override
-                public ExplainRoot asExplainRoot() {
-                    return this;
-                }
-
-                @Override
-                public String toString() {
-                    return "AnswerState.Partial.Compound.ExplainRoot{" +
-                            "root=" + root() +
-                            ", conceptMap=" + conceptMap() +
-                            ", requiresReiteration=" + requiresReiteration +
-                            ", explanation=" + explanation +
-                            '}';
-                }
-
-                @Override
-                public boolean equals(Object o) {
-                    if (this == o) return true;
-                    if (o == null || getClass() != o.getClass()) return false;
-                    ExplainRoot that = (ExplainRoot) o;
-                    return Objects.equals(root(), that.root()) &&
-                            Objects.equals(conceptMap, that.conceptMap) &&
-                            Objects.equals(parent, that.parent) &&
-                            Objects.equals(requiresReiteration, that.requiresReiteration) &&
-                            Objects.equals(explanation, that.explanation);
-                }
-
-                @Override
-                public int hashCode() {
-                    return hash;
                 }
 
             }
 
         }
 
-        public static abstract class Concludable<SLF extends Concludable<SLF, PRNT>, PRNT extends Compound<PRNT, ?>>
+        interface Concludable<SLF extends Concludable<SLF, PRNT>, PRNT extends Compound<PRNT, ?>> extends Partial<SLF, PRNT> {
+
+            Mapping mapping();
+
+            PRNT toUpstreamInferred();
+
+            Optional<? extends Conclusion<?, ?>> toDownstream(Unifier unifier, Rule rule);
+
+            @Override
+            default boolean isConcludable() { return true; }
+
+            @Override
+            default Concludable<?, ?> asConcludable() { return this; }
+
+            default boolean isMatch() { return false; }
+
+            default Match<?> asMatch() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Match.class); }
+
+            default boolean isExplain() { return false; }
+
+            default Explain asExplain() { throw GraknException.of(ILLEGAL_CAST, this.getClass(), Explain.class); }
+
+            interface Match<P extends Compound.Match<P, ?>> extends Concludable<Match<P>, P>, Explainable {
+
+                @Override
+                Optional<Conclusion.Match> toDownstream(Unifier unifier, Rule rule);
+
+                Match<P> with(ConceptMap extension, boolean requiresReiteration);
+
+                P toUpstreamLookup(ConceptMap additionalConcepts, boolean isInferredConclusion);
+
+                Conjunction conjunction();
+
+                @Override
+                default boolean isMatch() { return true; }
+
+                @Override
+                default Match<?> asMatch() { return this; }
+
+            }
+
+            interface Explain extends Concludable<Explain, Compound.Explain.Root> {
+
+                @Override
+                Optional<Conclusion.Explain> toDownstream(Unifier unifier, Rule rule);
+
+                Explain with(ConceptMap extension, boolean requiresReiteration, ConclusionAnswer conclusionAnswer);
+
+                ConclusionAnswer conclusionAnswer();
+
+                @Override
+                default boolean isExplain() {
+                    return true;
+                }
+
+                @Override
+                default Explain asExplain() { return this; }
+
+            }
+
+        }
+
+        interface Conclusion<SLF extends Conclusion<SLF, PRNT>, PRNT extends Concludable<?, ?>>
                 extends Partial<SLF, PRNT> {
 
-            final ConclusionAnswer conclusionAnswer; // TODO we only want this in the Explain
-            final Conjunction sourceConjunction; //TODO we only want this in Match, if at all since it's a bit odd
-            final Mapping mapping;
+            Rule rule();
 
-            public Concludable(ConceptMap partialAnswer, Mapping mapping, Conjunction sourceConjunction, @Nullable ConclusionAnswer conclusionAnswer,
-                               PRNT parent, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-                super(partialAnswer, parent, root, requiresReiteration);
-                this.mapping = mapping;
-                this.sourceConjunction = sourceConjunction;
-                this.conclusionAnswer = conclusionAnswer;
-            }
+            Unifier unifier();
 
-            static <P extends Compound.Match<P, ?>> Match<P> match(P parent, Mapping mapping, Actor.Driver<? extends Resolver<?>> root, Conjunction sourceConjunction, boolean explainable) {
-                ConceptMap mappedConceptMap = mapping.transform(parent.conceptMap());
-                return new Match<>(mappedConceptMap, parent, mapping, root, false, explainable, sourceConjunction, null);
-            }
+            Unifier.Requirements.Instance instanceRequirements();
 
-            static Explain explain(Compound.ExplainRoot parent, Mapping mapping, Actor.Driver<? extends Resolver<?>> root, Conjunction sourceConjunction) {
-                ConceptMap mappedConceptMap = mapping.transform(parent.conceptMap());
-                return new Explain(mappedConceptMap, parent, mapping, sourceConjunction, null, root, false);
-            }
+            Optional<? extends PRNT> aggregateToUpstream(Map<Identifier.Variable, Concept> concepts);
 
-            public abstract PRNT toUpstreamLookup(ConceptMap additionalConcepts, boolean isInferredConclusion);
 
-            public abstract PRNT toUpstreamInferred();
-
-            public abstract Optional<? extends Conclusion<?, ?>> toDownstream(Unifier unifier, Rule rule);
+            SLF extend(ConceptMap ans);
 
             @Override
-            public boolean isConcludable() { return true; }
+            default boolean isConclusion() { return true; }
 
             @Override
-            public Concludable<?, ?> asConcludable() { return this; }
+            default Conclusion<?, ?> asConclusion() { return this; }
 
-            public boolean isExplain() {
-                return false;
+            interface Match extends Conclusion<Match, Concludable.Match<?>>, Explainable {
+
+                @Override
+                Optional<Concludable.Match<?>> aggregateToUpstream(Map<Identifier.Variable, Concept> concepts);
+
+                Match with(ConceptMap extension, boolean requiresReiteration);
+
+                Compound.Match.Condition toDownstream(Set<Identifier.Variable.Retrievable> filter);
+
             }
 
-            public static class Match<P extends Compound.Match<P, ?>> extends Concludable<Match<P>, P> {
+            interface Explain extends Conclusion<Explain, Concludable.Explain> {
 
-                private final boolean explainable;
-                private final int hash;
+                ExplainableAnswer conditionAnswer();
 
-                private Match(ConceptMap mappedConceptMap, P parent, Mapping mapping, Actor.Driver<? extends Resolver<?>> root,
-                              boolean requiresReiteration, boolean explainable, Conjunction sourceConjunction, @Nullable ConclusionAnswer conclusionAnswer) {
-                    super(mappedConceptMap, mapping, sourceConjunction, conclusionAnswer, parent, root, requiresReiteration);
-                    this.explainable = explainable;
-                    this.hash = Objects.hash(root, conceptMap, mapping, parent, requiresReiteration, explainable);
-                }
+                Explain with(ConceptMap extension, boolean requiresReiteration, ExplainableAnswer conditionAnswer);
 
                 @Override
-                public P toUpstreamLookup(ConceptMap additionalConcepts, boolean isInferredConclusion) {
-                    if (isInferredConclusion) {
-                        return parent().with(mapping.unTransform(additionalConcepts), requiresReiteration || parent().requiresReiteration(),
-                                             sourceConjunction);
-                    } else {
-                        return parent().with(mapping.unTransform(additionalConcepts), requiresReiteration || parent().requiresReiteration());
-                    }
-                }
+                Optional<Concludable.Explain> aggregateToUpstream(Map<Identifier.Variable, Concept> concepts);
 
-                @Override
-                public P toUpstreamInferred() {
-                    return parent().with(mapping.unTransform(this.conceptMap()), requiresReiteration || parent().requiresReiteration(),
-                                         sourceConjunction);
-                }
+                Compound.Explain.Condition toDownstream(Set<Identifier.Variable.Retrievable> filter);
 
-                @Override
-                public Optional<Conclusion.Match> toDownstream(Unifier unifier, Rule rule) {
-                    return Conclusion.Match.match(this, unifier, rule, root(), explainable);
-                }
-
-                @Override
-                Concludable.Match<P> with(ConceptMap extension, boolean requiresReiteration) {
-                    // if we decide to keep an "explain" flag, we would use this endpoint
-                    throw GraknException.of(ILLEGAL_STATE);
-                }
-
-                Concludable.Match<P> with(ConceptMap extension, boolean requiresReiteration, ConclusionAnswer conclusionAnswer) {
-                    return new Concludable.Match<>(extendAnswer(extension), parent(), mapping, root(), requiresReiteration, explainable, sourceConjunction, conclusionAnswer);
-                }
-
-                @Override
-                public String toString() {
-                    return "AnswerState.Partial.Concludable.Match{" +
-                            "root=" + root() +
-                            ", conceptMap=" + conceptMap() +
-                            ", mapping=" + mapping +
-                            '}';
-                }
-
-                @Override
-                public boolean equals(Object o) {
-                    if (this == o) return true;
-                    if (o == null || getClass() != o.getClass()) return false;
-                    Match that = (Match) o;
-                    return Objects.equals(root(), that.root()) &&
-                            Objects.equals(conceptMap, that.conceptMap) &&
-                            Objects.equals(parent, that.parent) &&
-                            Objects.equals(mapping, that.mapping) &&
-                            requiresReiteration == that.requiresReiteration;
-                }
-
-                @Override
-                public int hashCode() {
-                    return hash;
-                }
             }
 
-            public static class Explain extends Concludable<Explain, Compound.ExplainRoot> {
-
-                private final int hash;
-
-                private Explain(ConceptMap mappedConceptMap, Compound.ExplainRoot parent, Mapping mapping, Conjunction sourceConjunction,
-                                @Nullable ConclusionAnswer conclusionAnswer, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-                    super(mappedConceptMap, mapping, sourceConjunction, conclusionAnswer, parent, root, requiresReiteration);
-                    //note: includes conclusion answer
-                    this.hash = Objects.hash(root, conceptMap, mapping, parent, requiresReiteration, conclusionAnswer);
-                }
-
-                @Override
-                public Compound.ExplainRoot toUpstreamLookup(ConceptMap additionalConcepts, boolean isInferredConclusion) {
-                    throw GraknException.of(ILLEGAL_STATE);
-//                    if (isInferredConclusion) {
-//                        return parent().with(mapping.unTransform(additionalConcepts), requiresReiteration || parent().requiresReiteration(), null);
-//                    } else {
-//                        return parent().with(mapping.unTransform(additionalConcepts), requiresReiteration || parent().requiresReiteration());
-//                    }
-                }
-
-                @Override
-                public Compound.ExplainRoot toUpstreamInferred() {
-                    Explanation explanation = new Explanation(conclusionAnswer.rule(), mapping, conclusionAnswer, conclusionAnswer.conditionAnswer());
-                    return parent().with(mapping.unTransform(this.conceptMap()), requiresReiteration || parent().requiresReiteration(), explanation);
-                }
-
-                @Override
-                public Optional<Conclusion.Explain> toDownstream(Unifier unifier, Rule rule) {
-                    return Conclusion.explain(this, unifier, rule, root());
-                }
-
-                @Override
-                Explain with(ConceptMap extension, boolean requiresReiteration) {
-                    // if we decide to keep an "explain" flag, we would use this endpoint
-                    throw GraknException.of(ILLEGAL_STATE);
-                }
-
-                Explain with(ConceptMap extension, boolean requiresReiteration, ConclusionAnswer conclusionAnswer) {
-                    return new Explain(extendAnswer(extension), parent(), mapping, sourceConjunction, conclusionAnswer, root(), requiresReiteration);
-                }
-
-                @Override
-                public boolean isExplain() {
-                    return true;
-                }
-
-                @Override
-                public String toString() {
-                    return "AnswerState.Partial.Concludable.Explain{" +
-                            "root=" + root() +
-                            ", conceptMap=" + conceptMap() +
-                            ", mapping=" + mapping +
-                            ", conclusionAnswer=" + conclusionAnswer +
-                            '}';
-                }
-
-                @Override
-                public boolean equals(Object o) {
-                    if (this == o) return true;
-                    if (o == null || getClass() != o.getClass()) return false;
-                    Explain that = (Explain) o;
-                    return Objects.equals(root(), that.root()) &&
-                            Objects.equals(conceptMap, that.conceptMap) &&
-                            Objects.equals(parent, that.parent) &&
-                            Objects.equals(mapping, that.mapping) &&
-                            Objects.equals(conclusionAnswer, that.conclusionAnswer) &&
-                            requiresReiteration == that.requiresReiteration;
-                }
-
-                @Override
-                public int hashCode() {
-                    return hash;
-                }
-            }
         }
 
-        public static abstract class Conclusion<SLF extends Conclusion<SLF, PRNT>, PRNT extends Concludable<?, ?>> extends Partial<SLF, PRNT> {
-
-            final Rule rule;
-            final Unifier unifier;
-            final Instance instanceRequirements;
-            final ExplainableAnswer conditionAnswer; // TODO we only want this in Explain, not in Match
-
-            public Conclusion(ConceptMap partialAnswer, PRNT parent, Rule rule, Unifier unifier,
-                              Instance instanceRequirements, @Nullable ExplainableAnswer conditionAnswer, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-                super(partialAnswer, parent, root, requiresReiteration);
-                this.rule = rule;
-                this.unifier = unifier;
-                this.instanceRequirements = instanceRequirements;
-                this.conditionAnswer = conditionAnswer;
-            }
-
-            static Optional<Match> match(Concludable.Match<?> parent, Unifier unifier, Rule rule, Actor.Driver<? extends Resolver<?>> root, boolean explainable) {
-                Optional<Pair<ConceptMap, Instance>> unified = unifier.unify(parent.conceptMap());
-                return unified.map(unification -> new Match(
-                        unification.first(), parent, rule, unifier, unification.second(), root, false, explainable, null
-                ));
-            }
-
-            static Optional<Explain> explain(Concludable.Explain parent, Unifier unifier, Rule rule, Actor.Driver<? extends Resolver<?>> root) {
-                Optional<Pair<ConceptMap, Instance>> unified = unifier.unify(parent.conceptMap());
-                return unified.map(unification -> new Explain(
-                        unification.first(), parent, rule, unifier, unification.second(), null, root, false
-                ));
-            }
-
-            abstract SLF with(ConceptMap extension, boolean requiresReiteration, ExplainableAnswer conditionAnswer);
-
-            // TODO it should be possible to generify this method implementation too, the generics get tricky
-            public abstract Optional<? extends PRNT> aggregateToUpstream(Map<Identifier.Variable, Concept> concepts);
-
-            // TODO it should be possible to generify this method implementation too, the generics get tricky
-            public abstract Compound.Match.Condition<SLF> toDownstream(Set<Identifier.Variable.Retrievable> filter);
-
-            public abstract SLF extend(ConceptMap ans);
-
-            @Override
-            public boolean isConclusion() { return true; }
-
-            @Override
-            public Conclusion<?, ?> asConclusion() { return this; }
-
-            ConceptMap toConceptMap(Map<Identifier.Variable, Concept> concepts) {
-                Map<Identifier.Variable.Retrievable, Concept> filteredMap = new HashMap<>();
-                iterate(concepts.entrySet()).filter(entry -> entry.getKey().isRetrievable())
-                        .forEachRemaining(entry -> filteredMap.put(entry.getKey().asRetrievable(), entry.getValue()));
-                return new ConceptMap(filteredMap);
-            }
-
-            public static class Match extends Conclusion<Match, Concludable.Match<?>> {
-
-                private final boolean explainable;
-                private final int hash;
-
-                private Match(ConceptMap unifiedConceptMap, Concludable.Match<?> parent, Rule rule, Unifier unifier,
-                              Instance instanceRequirements, Actor.Driver<? extends Resolver<?>> root,
-                              boolean requiresReiteration, boolean explainable, @Nullable ExplainableAnswer conditionAnswer) {
-                    super(unifiedConceptMap, parent, rule, unifier, instanceRequirements, conditionAnswer, root, requiresReiteration);
-                    this.explainable = explainable;
-                    this.hash = Objects.hash(root, conceptMap, rule, unifier, instanceRequirements, parent, requiresReiteration, explainable);
-                }
-
-                @Override
-                public Optional<Concludable.Match<?>> aggregateToUpstream(Map<Identifier.Variable, Concept> concepts) {
-                    Optional<ConceptMap> unUnified = unifier.unUnify(concepts, instanceRequirements);
-                    return unUnified.map(ans -> {
-                        // TODO we could make this "free" (object creation only) and skip the toConceptMap if we just store the raw map
-                        ConclusionAnswer conclusionAnswer = new ConclusionAnswer(rule, toConceptMap(concepts), unifier, conditionAnswer);
-                        return parent().with(ans, true, conclusionAnswer);
-                    });
-                }
-
-                @Override
-                public Compound.Match.Condition<Match> toDownstream(Set<Identifier.Variable.Retrievable> filter) {
-                    return Compound.Match.Condition.create(this, filter, root());
-                }
-
-                @Override
-                public Match extend(ConceptMap ans) {
-                    Map<Identifier.Variable.Retrievable, Concept> extended = new HashMap<>();
-                    extended.putAll(ans.concepts());
-                    extended.putAll(conceptMap.concepts());
-                    return new Match(new ConceptMap(extended), parent(), rule, unifier, instanceRequirements, root(), requiresReiteration, explainable, conditionAnswer
-                    );
-                }
-
-                Match with(ConceptMap extension, boolean requiresReiteration) {
-                    throw GraknException.of(ILLEGAL_STATE);
-                }
-
-                Match with(ConceptMap extension, boolean requiresReiteration, ExplainableAnswer conditionAnswer) {
-                    return new Match(extendAnswer(extension), parent(), rule, unifier, instanceRequirements, root(), requiresReiteration, explainable, conditionAnswer);
-                }
-
-                @Override
-                public String toString() {
-                    return "AnswerState.Partial.Conclusion.Match{" +
-                            "root=" + root() +
-                            ", conceptMap=" + conceptMap() +
-                            ", rule=" + rule.getLabel() +
-                            ", unifier=" + unifier +
-                            ", instanceRequirements=" + instanceRequirements +
-                            '}';
-                }
-
-                @Override
-                public boolean equals(Object o) {
-                    if (this == o) return true;
-                    if (o == null || getClass() != o.getClass()) return false;
-                    Match that = (Match) o;
-                    return Objects.equals(root(), that.root()) &&
-                            Objects.equals(conceptMap, that.conceptMap) &&
-                            Objects.equals(parent, that.parent) &&
-                            Objects.equals(rule, that.rule) &&
-                            Objects.equals(unifier, that.unifier) &&
-                            Objects.equals(instanceRequirements, that.instanceRequirements) &&
-                            requiresReiteration == that.requiresReiteration &&
-                            explainable == that.explainable;
-                }
-
-                @Override
-                public int hashCode() {
-                    return hash;
-                }
-            }
-
-            public static class Explain extends Conclusion<Explain, Concludable.Explain> {
-
-                private final int hash;
-
-                private Explain(ConceptMap unifiedConceptMap, Concludable.Explain parent, Rule rule, Unifier unifier,
-                                Instance instanceRequirements, @Nullable ExplainableAnswer conditionAnswer, Actor.Driver<? extends Resolver<?>> root, boolean requiresReiteration) {
-                    super(unifiedConceptMap, parent, rule, unifier, instanceRequirements, conditionAnswer, root, requiresReiteration);
-                    this.hash = Objects.hash(root, conceptMap, rule, unifier, instanceRequirements, parent, conditionAnswer);
-                }
-
-                @Override
-                public Optional<Concludable.Explain> aggregateToUpstream(Map<Identifier.Variable, Concept> concepts) {
-                    Optional<ConceptMap> unUnified = unifier.unUnify(concepts, instanceRequirements);
-                    return unUnified.map(ans -> {
-                        // TODO we could make this "free" (object creation only) and skip the toConceptMap if we just store the raw map
-                        ConclusionAnswer conclusionAnswer = new ConclusionAnswer(rule, toConceptMap(concepts), unifier, conditionAnswer);
-                        return parent().with(ans, true, conclusionAnswer);
-                    });
-                }
-
-                @Override
-                public Compound.Match.Condition<Explain> toDownstream(Set<Identifier.Variable.Retrievable> filter) {
-                    return Compound.Match.Condition.create(this, filter, root());
-                }
-
-                @Override
-                public Explain extend(ConceptMap ans) {
-                    Map<Identifier.Variable.Retrievable, Concept> extended = new HashMap<>();
-                    extended.putAll(ans.concepts());
-                    extended.putAll(conceptMap.concepts());
-                    return new Explain(new ConceptMap(extended), parent(), rule, unifier, instanceRequirements, conditionAnswer, root(),
-                                       requiresReiteration);
-                }
-
-                Explain with(ConceptMap extension, boolean requiresReiteration) {
-                    throw GraknException.of(ILLEGAL_STATE);
-                }
-
-                Explain with(ConceptMap extension, boolean requiresReiteration, ExplainableAnswer conditionAnswer) {
-                    return new Explain(extendAnswer(extension), parent(), rule, unifier, instanceRequirements, conditionAnswer, root(), requiresReiteration);
-                }
-
-                @Override
-                public String toString() {
-                    return "AnswerState.Partial.Conclusion{" +
-                            "root=" + root() +
-                            ", conceptMap=" + conceptMap() +
-                            ", rule=" + rule.getLabel() +
-                            ", unifier=" + unifier +
-                            ", instanceRequirements=" + instanceRequirements +
-                            ", conditionAnswer=" + conditionAnswer +
-                            '}';
-                }
-
-                @Override
-                public boolean equals(Object o) {
-                    if (this == o) return true;
-                    if (o == null || getClass() != o.getClass()) return false;
-                    Explain that = (Explain) o;
-                    return Objects.equals(root(), that.root()) &&
-                            Objects.equals(conceptMap, that.conceptMap) &&
-                            Objects.equals(parent, that.parent) &&
-                            Objects.equals(rule, that.rule) &&
-                            Objects.equals(unifier, that.unifier) &&
-                            Objects.equals(instanceRequirements, that.instanceRequirements) &&
-                            requiresReiteration == that.requiresReiteration &&
-                            Objects.equals(conditionAnswer, that.conditionAnswer);
-                }
-
-                @Override
-                public int hashCode() {
-                    return hash;
-                }
-            }
-        }
     }
+
 }
+
