+++ reasoner/resolution/framework/Request.java
@@ -19,7 +19,7 @@
 package grakn.core.reasoner.resolution.framework;
 
 import grakn.core.concurrent.actor.Actor;
-import grakn.core.reasoner.resolution.answer.AnswerState.Partial;
+import grakn.core.reasoner.resolution.answer.AnswerStateOld.Partial;
 
 import javax.annotation.Nullable;
 import java.util.Objects;
